generators/product_qr/requirements.txt
pyyaml
qrcode[pil]